Deprecate the org.http4s.util.execution package
