+++ servlet/src/main/scala/org/http4s/servlet/ServletIo.scala
@@ -7,8 +7,8 @@ import fs2._
 import java.util.concurrent.atomic.AtomicReference
 import javax.servlet.{ReadListener, WriteListener}
 import javax.servlet.http.{HttpServletRequest, HttpServletResponse}
+import org.http4s.internal.Trampoline
 import org.http4s.util.bug
-import org.http4s.util.execution.trampoline
 import org.log4s.getLogger
 import scala.annotation.tailrec
 
@@ -100,7 +100,7 @@ final case class NonBlockingServletIo[F[_]: Effect](chunkSize: Int) extends Serv
         // This effect sets the callback and waits for the first bytes to read
         val registerRead =
           // Shift execution to a different EC
-          Async.shift(trampoline) *>
+          Async.shift(Trampoline) *>
             F.async[Option[Chunk[Byte]]] { cb =>
               if (!state.compareAndSet(Init, Blocked(cb))) {
                 cb(Left(bug("Shouldn't have gotten here: I should be the first to set a state")))
@@ -131,7 +131,7 @@ final case class NonBlockingServletIo[F[_]: Effect](chunkSize: Int) extends Serv
         val readStream = Stream.eval(registerRead) ++ Stream
           .repeatEval( // perform the initial set then transition into normal read mode
             // Shift execution to a different EC
-            Async.shift(trampoline) *>
+            Async.shift(Trampoline) *>
               F.async[Option[Chunk[Byte]]] { cb =>
                 @tailrec
                 def go(): Unit = state.get match {
@@ -221,7 +221,7 @@ final case class NonBlockingServletIo[F[_]: Effect](chunkSize: Int) extends Serv
 
     val awaitLastWrite = Stream.eval_ {
       // Shift execution to a different EC
-      Async.shift(trampoline) *>
+      Async.shift(Trampoline) *>
         F.async[Unit] { cb =>
           state.getAndSet(AwaitingLastWrite(cb)) match {
             case Ready if out.isReady => cb(Right(()))
@@ -236,7 +236,7 @@ final case class NonBlockingServletIo[F[_]: Effect](chunkSize: Int) extends Serv
       response.body.chunks
         .evalMap { chunk =>
           // Shift execution to a different EC
-          Async.shift(trampoline) *>
+          Async.shift(Trampoline) *>
             F.async[Chunk[Byte] => Unit] { cb =>
                 val blocked = Blocked(cb)
                 state.getAndSet(blocked) match {
