+++ tests/src/test/scala/org/http4s/EntityDecoderSpec.scala
@@ -10,13 +10,13 @@ import java.nio.charset.StandardCharsets
 import cats.data.Chain
 import org.http4s.Status.Ok
 import org.http4s.headers.`Content-Type`
+import org.http4s.internal.Trampoline
 import org.http4s.testing.Http4sLegacyMatchersIO
-import org.http4s.util.execution.trampoline
 import org.specs2.execute.PendingUntilFixed
 import scala.concurrent.ExecutionContext
 
 class EntityDecoderSpec extends Http4sSpec with Http4sLegacyMatchersIO with PendingUntilFixed {
-  implicit val executionContext: ExecutionContext = trampoline
+  implicit val executionContext: ExecutionContext = Trampoline
   implicit val testContext: TestContext = TestContext()
 
   val `application/excel`: MediaType =
